# sez/src/main/resources/api/swagger_full.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "API for managing the tutoring system (CRM)",
    "title" : "SEZ Tutoring Platform API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "Local Server",
    "url" : "http://localhost:8080"
  } ],
  "paths" : {
...
    "/api/v1/courses" : {
      "get" : {
        "operationId" : "searchCourses",
        "parameters" : [ {
          "description" : "Filter by tutor first or last name",
          "in" : "query",
          "name" : "tutorName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Filter by course (subject) name",
          "in" : "query",
          "name" : "courseName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Maximum hourly rate",
          "in" : "query",
          "name" : "maxRate",
          "schema" : {
            "type" : "number"
          }
        }, {
          "description" : "If true, returns only courses where the current student is not enrolled",
          "in" : "query",
          "name" : "excludeEnrolled",
          "schema" : {
            "default" : true,
            "type" : "boolean"
          }
        }, {
          "description" : "Sorting criteria. Multiple values allowed (e.g., sortBy=rating_desc&sortBy=price_asc)",
          "explode" : true,
          "in" : "query",
          "name" : "sortBy",
          "schema" : {
            "items" : {
              "enum" : [ "price_asc", "price_desc", "rating_asc", "rating_desc" ],
              "type" : "string"
            },
            "type" : "array"
          },
          "style" : "form"
        }, {
          "description" : "Page number",
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "Number of items per page (max 20)",
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "maximum" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CourseResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Successful response"
          }
        },
        "summary" : "Search available courses with filters for students",
        "tags" : [ "Courses" ]
      },
      "post" : {
        "operationId" : "createCourse",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCourseDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Course created successfully"
          }
        },
        "summary" : "Create a new course offering",
        "tags" : [ "Courses" ]
      }
    },
    "/api/v1/courses/my" : {
      "get" : {
        "operationId" : "getMyCourses",
        "parameters" : [ {
          "description" : "Filter by course (subject) name",
          "in" : "query",
          "name" : "courseName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "maximum" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CourseResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of personal courses"
          }
        },
        "summary" : "Get personal courses for tutor",
        "tags" : [ "Courses" ]
      }
    },
    "/api/v1/courses/{id}" : {
      "delete" : {
        "description" : "Deletion will be blocked if active enrollments exist",
        "operationId" : "deleteCourse",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Course deleted successfully"
          },
          "400" : {
            "description" : "Cannot delete course with active students"
          },
          "404" : {
            "description" : "Course not found"
          }
        },
        "summary" : "Delete course by ID",
        "tags" : [ "Courses" ]
      },
      "get" : {
        "operationId" : "getCourseById",
        "parameters" : [ {
          "description" : "Unique identifier of the course",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CourseResponse"
                }
              }
            },
            "description" : "Detailed information about the course"
          },
          "404" : {
            "description" : "Course not found"
          }
        },
        "summary" : "Get course details by ID",
        "tags" : [ "Courses" ]
      },
      "patch" : {
        "operationId" : "updateCourse",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateCourseDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Course updated successfully"
          },
          "404" : {
            "description" : "Course not found"
          }
        },
        "summary" : "Update course details",
        "tags" : [ "Courses" ]
      }
    },
    "/api/v1/enrollments" : {
      "get" : {
        "operationId" : "getEnrollments",
        "parameters" : [ {
          "description" : "Filter by student name or surname",
          "in" : "query",
          "name" : "studentName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Filter by tutor name or surname",
          "in" : "query",
          "name" : "tutorName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Filter by subject/course name",
          "in" : "query",
          "name" : "courseName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Page number",
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "Number of items per page (max 20)",
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 20,
            "maximum" : 20,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EnrollmentResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of enrollments retrieved"
          }
        },
        "summary" : "Get list of enrollments for student or tutor",
        "tags" : [ "Enrollments" ]
      },
      "post" : {
        "operationId" : "requestEnrollment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEnrollmentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Enrollment request sent successfully"
          }
        },
        "summary" : "Request enrollment in a course",
        "tags" : [ "Enrollments" ]
      }
    },
    "/api/v1/enrollments/{id}" : {
      "delete" : {
        "description" : "Deletes enrollment and associated planned lessons",
        "operationId" : "terminateEnrollment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Enrollment terminated successfully"
          },
          "404" : {
            "description" : "Enrollment not found"
          }
        },
        "summary" : "Terminate enrollment",
        "tags" : [ "Enrollments" ]
      },
      "patch" : {
        "operationId" : "updateEnrollmentStatus",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateEnrollmentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Enrollment status updated successfully"
          },
          "404" : {
            "description" : "Enrollment not found"
          }
        },
        "summary" : "Confirm or update enrollment status",
        "tags" : [ "Enrollments" ]
      }
    },
    "/api/v1/enrollments/{id}/details" : {
      "get" : {
        "operationId" : "getEnrollmentDetails",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnrollmentDetailsResponse"
                }
              }
            },
            "description" : "Full enrollment details retrieved"
          },
          "404" : {
            "description" : "Enrollment not found"
          }
        },
        "summary" : "Get full enrollment details including student and course objects",
        "tags" : [ "Enrollments" ]
      }
    },
    "/api/v1/enrollments/my-active" : {
      "get" : {
        "operationId" : "getMyActiveEnrollments",
        "parameters" : [ {
          "description" : "Filter by student name or surname",
          "in" : "query",
          "name" : "studentName",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Filter by subject/course name",
          "in" : "query",
          "name" : "courseName",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EnrollmentDetailsResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "List of active enrollments retrieved"
          }
        },
        "summary" : "Get list of active enrollments for current student",
        "tags" : [ "Enrollments" ]
      }
    },
    "/api/v1/enrollments/by-course/{courseId}" : {
      "delete" : {
        "description" : "Видаляє запис на навчання за ID курсу.  ID поточного студента визначається автоматично сервером через токен авторизації.\n",
        "operationId" : "terminateEnrollmentByCourse",
        "parameters" : [ {
          "description" : "ID курсу, запис на який потрібно видалити",
          "in" : "path",
          "name" : "courseId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Enrollment terminated successfully"
          },
          "401" : {
            "description" : "Unauthorized (користувач не авторизований)"
          },
          "404" : {
            "description" : "Enrollment not found for this course and current student"
          }
        },
        "summary" : "Terminate enrollment by course ID",
        "tags" : [ "Enrollments" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "StudentResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "role" : {
            "enum" : [ "STUDENT" ],
            "type" : "string"
          }
        },
        "required" : [ "email", "id", "name", "role", "surname" ],
        "type" : "object"
      },
...
      "CourseResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "tutorName" : {
            "type" : "string"
          },
          "tutorSurname" : {
            "type" : "string"
          },
          "tutorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "subjectName" : {
            "type" : "string"
          },
          "hourlyRate" : {
            "type" : "number"
          },
          "description" : {
            "type" : "string"
          },
          "rating" : {
            "maximum" : 5,
            "minimum" : 1,
            "type" : "number"
          }
        },
        "required" : [ "description", "hourlyRate", "id", "rating", "subjectName", "tutorId", "tutorName", "tutorSurname" ],
        "type" : "object"
      },
      "CreateCourseDto" : {
        "properties" : {
          "subjectId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "hourlyRate" : {
            "description" : "Hourly rate for tutoring services. Must be at least 1.0",
            "minimum" : 1,
            "type" : "number"
          },
          "description" : {
            "description" : "Course description for potential students",
            "maxLength" : 1000,
            "minLength" : 2,
            "type" : "string"
          }
        },
        "required" : [ "description", "hourlyRate", "subjectId" ],
        "type" : "object"
      },
      "UpdateCourseDto" : {
        "properties" : {
          "hourlyRate" : {
            "description" : "Hourly rate for tutoring services. Must be at least 1.0",
            "minimum" : 1,
            "type" : "number"
          },
          "description" : {
            "description" : "Course description for potential students",
            "maxLength" : 1000,
            "minLength" : 2,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "EnrollmentResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "courseId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "subjectName" : {
            "type" : "string"
          },
          "tutorName" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "REQUESTED", "ACTIVE" ],
            "type" : "string"
          }
        },
        "required" : [ "courseId", "id", "status", "studentId", "subjectName", "tutorName" ],
        "type" : "object"
      },
      "CreateEnrollmentDto" : {
        "properties" : {
          "courseId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "courseId" ],
        "type" : "object"
      },
      "UpdateEnrollmentDto" : {
        "properties" : {
          "status" : {
            "description" : "Use ACTIVE to confirm the enrollment request",
            "enum" : [ "ACTIVE" ],
            "type" : "string"
          }
        },
        "required" : [ "status" ],
        "type" : "object"
      },
      "EnrollmentDetailsResponse" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "REQUESTED", "ACTIVE" ],
            "type" : "string"
          },
          "student" : {
            "$ref" : "#/components/schemas/StudentResponse"
          },
          "course" : {
            "$ref" : "#/components/schemas/CourseResponse"
          }
        },
        "required" : [ "course", "id", "status", "student" ],
        "type" : "object"
      },
...
    }
  }
}